Split `SHs.AST.Expr` out of `SHs.AST`
Extract expression and binding types into `HsBindgen.Backend.SHs.AST.Expr`

Remove unused `VarDeclRHS` and `VarDeclRHSAppHead` types from `Hs.AST`.

Reduce dependencies on `c-expr-dsl` (`canBeRepresentedAsRational`).
