+++ hs-bindgen/src-internal/HsBindgen/Backend/SHs/AST.hs
@@ -36,11 +36,6 @@ module HsBindgen.Backend.SHs.AST (
   , PatternSynonym(..)
   ) where
 
-import Data.Type.Nat (Nat1)
-import DeBruijn (Add, Ctx, EmptyCtx, Idx)
-
-import C.Char qualified as CExpr.Runtime
-
 import HsBindgen.Backend.Global
 import HsBindgen.Backend.Hs.AST.Strategy qualified as Hs
 import HsBindgen.Backend.Hs.CallConv
@@ -48,110 +43,25 @@ import HsBindgen.Backend.Hs.Haddock.Documentation qualified as HsDoc
 import HsBindgen.Backend.Hs.Name qualified as Hs
 import HsBindgen.Backend.Hs.Origin qualified as Origin
 import HsBindgen.Backend.Level
+import HsBindgen.Backend.SHs.AST.Expr
 import HsBindgen.Backend.SHs.AST.Type
 import HsBindgen.Frontend.Naming
 import HsBindgen.Imports
 import HsBindgen.Instances qualified as Inst
 import HsBindgen.Language.Haskell qualified as Hs
-import HsBindgen.NameHint
+
+{-------------------------------------------------------------------------------
+  Re-exports from SHs.AST.Expr and SHs.AST.Type
+-------------------------------------------------------------------------------}
+
+-- SExpr, ClosedExpr, SAlt, PatExpr, InfixOp, Binding, Parameter, Result,
+-- Pragma are re-exported from SHs.AST.Expr (imported above).
+-- SType, ClosedType, Plus2, applyPlus2, tBindgenGlobal are re-exported from
+-- SHs.AST.Type (imported above).
 
 {-------------------------------------------------------------------------------
   Backend representation
 -------------------------------------------------------------------------------}
-
-type ClosedExpr = SExpr EmptyCtx
-
--- | Simple expressions
-type SExpr :: Ctx -> Star
-data SExpr ctx =
-    EGlobal (Global LvlTerm)
-  | EBound (Idx ctx)
-  | EFree Hs.TermName
-  | ECon (Hs.Name Hs.NsConstr)
-  | EUnboxedIntegral Integer
-  | EIntegral Integer (Maybe ClosedType)
-  | EFloat Float ClosedType   -- ^ Type annotation to distinguish Float/CFloat
-  | EDouble Double ClosedType -- ^ Type annotation to distinguish Double/CDouble
-  | EChar CExpr.Runtime.CharValue
-  | EString String
-  | ECString ByteArray
-  | EApp (SExpr ctx) (SExpr ctx)
-  | EInfix InfixOp (SExpr ctx) (SExpr ctx)
-  | ELam NameHint (SExpr (S ctx))
-  | EUnusedLam (SExpr ctx)
-  | ECase (SExpr ctx) [SAlt ctx]
-  | EUnit
-  | EBoxedTup Plus2
-  | EUnboxedTup Plus2
-  | EList [SExpr ctx]
-    -- | Type application using \@
-  | ETypeApp (SExpr ctx) ClosedType
-  deriving stock (Show)
-
-eBindgenGlobal :: BindgenGlobalTerm -> SExpr ctx
-eBindgenGlobal = EGlobal . bindgenGlobalTerm
-
--- | Pattern&Expressions
---
--- There is common sublanguage between term and pattern expressions.
--- This is that common sublanguage.
--- Expressions ('SExpr') and full patterns (may) have more cases,
--- but for bidirectional pattern synonyms only the common bases
--- can be used.
---
--- For now 'PatExpr' is quite small, as we don't need much.
-type PatExpr :: Star
-data PatExpr
-  = PEApps (Hs.Name Hs.NsConstr) [PatExpr] -- head of pattern application cannot be variable.
-  | PELit Integer
-  deriving stock (Show)
-
-eInt :: Int -> SExpr be
-eInt i = EIntegral (fromIntegral i) (Just $ tBindgenGlobal Int_type)
-
--- | Supported infix operators.
-data InfixOp =
-    InfixApplicative_seq
-  | InfixMonad_seq
-  | InfixNonEmpty_constructor
-  deriving stock (Show, Eq)
-
-infixOpGlobal :: InfixOp -> Global LvlTerm
-infixOpGlobal = bindgenGlobalTerm . \case
-  InfixApplicative_seq      -> Applicative_seq
-  InfixMonad_seq            -> Monad_seq
-  InfixNonEmpty_constructor -> NonEmpty_constructor
-
--- | Case alternatives
---
--- Represents different kinds of pattern matching alternatives in case expressions.
---
--- Examples:
---
--- > case x of
--- >   Just y -> ...     -- SAlt with constructor "Just"
--- >   5# -> ...         -- SAltNoConstr for primitive literal
--- >   (# a, b #) -> ... -- SAltUnboxedTuple
---
-data SAlt ctx where
-    -- | Constructor pattern: @Con x y z -> body@
-    --
-    -- Pattern matches against a data constructor with fields.
-    SAlt
-      :: Hs.Name Hs.NsConstr -> Add n ctx ctx' -> Vec n NameHint -> SExpr ctx' -> SAlt ctx
-    -- | Non-constructor pattern: @5# -> body@ or @x -> body@
-    --
-    -- Used for matching primitive values, literals, or catch-all variables.
-    -- Always binds exactly one variable (hence Vec Nat1).
-    SAltNoConstr
-      :: Vec Nat1 NameHint -> SExpr (S ctx) -> SAlt ctx
-    -- | Unboxed tuple pattern: @(# x, y #) -> body@
-    --
-    -- Pattern matches against unboxed tuples.
-    SAltUnboxedTuple
-      :: Add n ctx ctx' -> Vec n NameHint -> SExpr ctx' -> SAlt ctx
-
-deriving stock instance Show (SAlt ctx)
 
 -- | Simple declarations
 data SDecl =
@@ -165,12 +75,6 @@ data SDecl =
   | DBinding Binding
   | DPatternSynonym PatternSynonym
   deriving stock (Show)
-
-tBindgenGlobal :: BindgenGlobalType -> SType ctx
-tBindgenGlobal = TGlobal . bindgenGlobalType
-
-data Pragma = NOINLINE
-  deriving stock Show
 
 data TypeSynonym = TypeSynonym{
       name    :: Hs.Name Hs.NsTypeConstr
@@ -232,10 +136,6 @@ data Newtype = Newtype{
     }
   deriving stock (Show, Generic)
 
--- | We might want to reconsider the decision of 'HsBindgen.Backend.Hs.AST.foreignImportParameters' as
--- well as 'HsBindgen.Backend.Hs.AST.foreignImportResultType' being 'ClosedType's if we ever want to
--- generate polymorphic type signatures.
---
 data ForeignImport = ForeignImport{
       name       :: Hs.TermName
     , parameters :: [Parameter]
@@ -254,28 +154,6 @@ data Safety = Safe | Unsafe
 
 instance Default Safety where
   def = Safe
-
-data Binding = Binding{
-      name       :: Hs.TermName
-    , parameters :: [Parameter]
-    , result     :: Result
-    , body       :: ClosedExpr
-    , pragmas    :: [Pragma]
-    , comment    :: Maybe HsDoc.Comment
-    }
-  deriving stock (Show, Generic)
-
-data Parameter = Parameter{
-      typ     :: ClosedType
-    , comment :: Maybe HsDoc.Comment
-    }
-  deriving stock (Show, Generic)
-
-data Result = Result{
-      typ     :: ClosedType
-    , comment :: Maybe HsDoc.Comment
-    }
-  deriving stock (Show, Generic)
 
 data PatternSynonym = PatternSynonym{
       name    :: Hs.Name Hs.NsConstr
